tint/resolver: Temporally ban f16 in buffer, pipeline IO and override
This patch make resolver reject using f16 types in uniform or storage
buffer, pipeline IO or overridable variable, since these are not
implemented yet. This can help prevent hitting invalid path in writers.

Bug: tint:1473, tint:1502
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/95827
Auto-Submit: Zhaoming Jiang <zhaoming.jiang@intel.com>
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Zhaoming Jiang <zhaoming.jiang@intel.com>

+++ src/tint/resolver/host_shareable_validation_test.cc
@@ -107,7 +107,7 @@ TEST_F(ResolverHostShareableValidationTest, NoError) {
     Enable(ast::Extension::kF16);
 
     auto* i1 = Structure("I1", {
-                                   Member(Source{{1, 1}}, "w1", ty.f16()),
+                                   Member(Source{{1, 1}}, "w1", ty.f32()),
                                    Member(Source{{2, 1}}, "x1", ty.f32()),
                                    Member(Source{{3, 1}}, "y1", ty.vec3<f32>()),
                                    Member(Source{{4, 1}}, "z1", ty.array<i32, 4>()),
@@ -115,7 +115,7 @@ TEST_F(ResolverHostShareableValidationTest, NoError) {
     auto* a1 = Alias("a1", ty.Of(i1));
     auto* i2 = Structure("I2", {
                                    Member(Source{{5, 1}}, "x2", ty.mat2x2<f32>()),
-                                   Member(Source{{6, 1}}, "w2", ty.mat3x4<f16>()),
+                                   Member(Source{{6, 1}}, "w2", ty.mat3x4<f32>()),
                                    Member(Source{{7, 1}}, "z2", ty.Of(i1)),
                                });
     auto* a2 = Alias("a2", ty.Of(i2));
